Moved FCM to Android exclusive
